Bump vertx
