+++ adoptopenjdk-frontend-parent/adoptopenjdk-api-v3-frontend/src/test/kotlin/net/adoptopenjdk/api/V3Test.kt
@@ -1,6 +1,7 @@
 package net.adoptopenjdk.api
 
-import net.adoptopenjdk.api.v3.V3
+import net.adoptopenjdk.api.v3.Startup
+import net.adoptopenjdk.api.v3.Startup.Companion.ENABLE_PERIODIC_UPDATES
 import net.adoptopenjdk.api.v3.dataSources.APIDataStore
 import org.junit.jupiter.api.extension.ExtendWith
 import kotlin.test.Test
@@ -11,8 +12,8 @@ class V3Test : FrontendTest() {
 
     @Test
     fun `update is scheduled`(apiDataStore: APIDataStore) {
-        System.setProperty(V3.ENABLE_PERIODIC_UPDATES, "true")
-        V3().schedulePeriodicUpdates(apiDataStore)
+        System.setProperty(ENABLE_PERIODIC_UPDATES, "true")
+        Startup(apiDataStore).schedulePeriodicUpdates()
         assertNotNull((apiDataStore as ApiDataStoreStub).scheduled)
     }
 }
