+++ adoptopenjdk-updater-parent/adoptopenjdk-api-v3-updater/pom.xml
@@ -194,7 +194,6 @@
                     <plugin>
                         <groupId>org.apache.maven.plugins</groupId>
                         <artifactId>maven-surefire-plugin</artifactId>
-                        <version>${surefire-plugin.version}</version>
                         <configuration>
                             <excludes>
                                 <exclude>**/StatsCalculatorTest.java</exclude>
